scripts/requirements.txt
py-solc-x==2.0.3
web3==5.31.1
argparse==1.4.0
